requirements.txt
openpyxl
pandas~=2.3.3
numpy~=2.3.5
pywebview~=6.1